Extract types and utils
