+++ src/types/SearchResult.d.ts
@@ -1,0 +1,9 @@
+export type SearchResult = {
+  name: string;
+  html_url: string;
+  stargazers_count: number;
+  owner: {
+    login: string;
+    avatar_url: string;
+  };
+};
